refactor: extract dropdown from TanstackTable to ApplicationTable specifically
