+++ app/page.tsx
@@ -1,10 +1,6 @@
 import ApplicationTable from '@/components/ApplicationTable'
-import FilterDropdown from '@/components/TanstackTable/FilterDropdown'
 import prisma from '@/db'
 import { Role } from '@prisma/client'
-import { Card } from '@radix-ui/themes'
-
-import FilterBar from './FilterBar'
 
 export default async function Home() {
   const applications = await prisma.application.findMany({
