+++ components/ApplicationTable.tsx
@@ -1,11 +1,13 @@
 'use client'
 
-import FilterBar from '@/app/FilterBar'
-import { Applicant, Application, User } from '@prisma/client'
-import { createColumnHelper } from '@tanstack/react-table'
-import React, { FC } from 'react'
+import { Applicant, Application, NextAction, User } from '@prisma/client'
+import { Card } from '@radix-ui/themes'
+import { ColumnFiltersState, createColumnHelper } from '@tanstack/react-table'
+import { usePathname, useRouter, useSearchParams } from 'next/navigation'
+import React, { FC, useEffect, useState } from 'react'
 
 import TanstackTable from './TanstackTable'
+import FilterDropdown from './TanstackTable/FilterDropdown'
 
 type ApplicationRow = Pick<Application, 'nextAction' | 'feeStatus' | 'wideningParticipation'> & {
   applicant: Pick<Applicant, 'cid' | 'ucasNumber' | 'firstName' | 'surname'>
@@ -62,11 +64,66 @@ interface ApplicationTableProps {
   reviewerIds: number[]
 }
 
-const ApplicationTable: FC<ApplicationTableProps> = ({ applications }) => (
-  <>
-    <FilterBar />
-    <TanstackTable data={applications} columns={columns} />
-  </>
-)
+const ApplicationTable: FC<ApplicationTableProps> = ({ applications }) => {
+  const router = useRouter()
+  const pathname = usePathname()
+  const searchParams = useSearchParams()
+  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([])
+  const [nextActionFilterValue, setNextActionFilterValue] = useState('ALL')
+
+  // searchParams determine what filters should be applied and the value of the dropdown
+  useEffect(() => {
+    setColumnFilters(
+      searchParams.get('nextAction')
+        ? [{ id: 'nextAction', value: searchParams.get('nextAction') }]
+        : []
+    )
+    setNextActionFilterValue(searchParams.get('nextAction') || 'ALL')
+  }, [searchParams, setNextActionFilterValue, setColumnFilters])
+
+  const updateSearchParam = (name: string, value: string) => {
+    const params = new URLSearchParams(searchParams.toString())
+    params.set(name, value)
+    router.push(pathname + '?' + params.toString())
+  }
+
+  const removeSearchParam = (name: string) => {
+    const params = new URLSearchParams(searchParams.toString())
+    params.delete(name)
+    router.push(pathname + '?' + params.toString())
+  }
+
+  // update searchParams which in turn update the dropdown value and the filters
+  const onFilterDropdownChange = (name: string, value: string) => {
+    if (value === 'ALL') {
+      removeSearchParam(name)
+    } else {
+      updateSearchParam(name, value)
+    }
+  }
+
+  return (
+    <>
+      <Card>
+        <FilterDropdown
+          values={['apple', 'banana']}
+          currentValue={'apple'}
+          onValueChange={() => {}}
+        />
+        <FilterDropdown
+          onValueChange={(value) => onFilterDropdownChange('nextAction', value)}
+          values={[...Object.keys(NextAction), 'ALL']}
+          currentValue={nextActionFilterValue}
+        />
+      </Card>
+      <TanstackTable
+        data={applications}
+        columns={columns}
+        columnFilters={columnFilters}
+        setColumnFilters={setColumnFilters}
+      />
+    </>
+  )
+}
 
 export default ApplicationTable
